SEQUOIADBMAINSTREAM-3880 sac 1.修复MySQL创建表，字段类型为char/varchar时，长度选项填写数值过大，报错提示错误的问题。
git-svn-id: http://192.168.20.11/sequoiadb/trunk@37274 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ tools/om_plugins/sequoiasql/www/app/controller/Data/MySQL/Operate/Structure.js
@@ -135,7 +135,7 @@
                            "name": "length",
                            "webName": $scope.pAutoLanguage( "长度" ),
                            "placeholder": $scope.pAutoLanguage( "长度" ),
-                           "type": "int",
+                           "type": "double",
                            "value": "",
                            "valid": {
                               "min": 0,
